pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ctracing"
dynamic = ["version"]
authors = [{ name = "Damien Degois", email = "damien@degois.info" }]
description = "custom tracing lib and helpers"
dependencies = ["opentelemetry-api", "opentelemetry-sdk"]
classifiers = [
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Programming Language :: Python :: 3 :: Only"
  ]
[project.optional-dependencies]
http = ["requests"]
otlp = ["opentelemetry-exporter-otlp"]
full = ["ctracing[http,otlp]"]
tests = ["ctracing[full]", "pytest"]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.dynamic]
version = { file = "VERSION" }

[tool.pytest.ini_options]
addopts = "--cov=ctracing"
testpaths = ["tests"]
filterwarnings = ["ignore:Call to deprecated method __init__."]


[tool.mypy]
mypy_path = "src"
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true
